Extract some functions into a helper
Why:
Some functions maybe needed by some other modules

This change addresses the need by:
Extract some general functions into a separate
`helper.js` file
