+++ src/lib/query-parser.js
@@ -1,31 +1,18 @@
 import R from 'ramda';
+import {
+  isNotEmpty,
+  isObject,
+  getFirstKey,
+  getFirstValue,
+  executeOrThrowWhenNil,
+  stringToArray,
+} from './helper';
 
-const isNotEmpty = R.compose(R.not, R.isEmpty);
-const isObject = R.allPass([R.is(Object), R.compose(R.not, R.isArrayLike)]);
-const isSplitable = R.allPass([R.is(String), R.test(/,/)]);
-const getFirstKey = R.pipe(R.keys, R.head);
-const getFirstValue = R.pipe(R.values, R.head);
 const isNotRelation = R.allPass([R.is(String), R.compose(R.not, R.test(/\./))]);
 const isRelation = R.allPass([R.is(String), R.test(/\./)]);
 const throwUnsupportedOperator = (o) => {
   throw new Error(`Unsuppported operator: ${o}`);
 };
-const throwError = R.curry((fnError, params) =>
-  () => fnError(...params));
-const executeOrThrowWhenNil = R.curry((fnError, errParams, fnParams) =>
-  R.ifElse(
-    R.isNil,
-    throwError(fnError, errParams),
-    f => f(...fnParams),
-  ));
-
-
-const stringToArray = R.cond([
-  [R.isEmpty, R.empty],
-  [isSplitable, R.compose(R.map(R.trim), R.split(','))],
-  [R.is(String), R.of],
-  [R.T, R.identity],
-]);
 
 const buildParserFunction = R.curry((fn, key, q, o) =>
   R.pipe(
